warehouse_management/requirements.txt
# Core dependencies
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pydantic==2.4.2
pandas==2.1.1
numpy==1.26.0
scipy==1.11.3

# Visualization
plotly==5.18.0
matplotlib==3.8.0
folium==0.14.0
kaleido==0.2.1  # For static image export with plotly

# Database
geoalchemy2==0.14.0
spatialite==0.0.3

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
tabulate==0.9.0
colorama==0.4.6
